fix(team): push-interval polish — copy, error surfacing, single TeamConfig, admin gate consistency
